backend/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.25
geoalchemy2>=0.14.3
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Data Processing
pydantic>=2.10.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0

# GIS
shapely>=2.0.2
geojson>=3.1.0
pyproj>=3.6.1

# Name Similarity (Hindi/English)
python-Levenshtein>=0.23.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-multipart>=0.0.6
httpx>=0.26.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3

